PrettyDoc

Documentation for PrettyDoc.

PrettyDoc.comma

PrettyDoc.defaultRenderOptions

PrettyDoc.empty

PrettyDoc.space

PrettyDoc.Doc

PrettyDoc.RenderOptions

PrettyDoc.align

PrettyDoc.angle

PrettyDoc.brace

PrettyDoc.bracket

PrettyDoc.breakable

PrettyDoc.indent

PrettyDoc.listof

PrettyDoc.parens

PrettyDoc.pretty

PrettyDoc.seg

PrettyDoc.seplistof

PrettyDoc.vsep

PrettyDoc.comma — Constant

A document representing a single comma.

PrettyDoc.defaultRenderOptions — Constant

A default set of rendering options.

PrettyDoc.empty — Constant

A empty document.

PrettyDoc.space — Constant

A document object representing a single whitespace.

PrettyDoc.Doc — Type

Combinatorial documentation objects, which has the following constructors:

Doc_Concat(left::Doc, right::Doc) Concatenates two documents,and the line is not broken.

Doc_VSep(elements::Vector{Doc}) Concatenates a vector of documents with the line separator `"

"`.

Doc_Align(document::Doc) Aligns the document using the current indentation.

Doc_Indent(delta::Int, document::Doc) Locally indents the document by delta spaces.

Doc_Segment(segment::String) A segment of text.

Doc_Breakable(document::Doc) Inidicating the end of the document is breakable.

Doc_Empty Do nothing.

PrettyDoc.RenderOptions — Type

Rendering options

PrettyDoc.align — Method

Align the document using the current indentation.

PrettyDoc.angle — Method

Wrap a document with left and right angle brackets.

PrettyDoc.brace — Method

Wrap a document with left and right braces.

PrettyDoc.bracket — Method

Wrap a document with left and right brackets.

PrettyDoc.breakable — Method

Make the document breakable at its beginning.

PrettyDoc.indent — Method

Indent a document by delta spaces.

PrettyDoc.listof — Function

Create a document that represents a list of elements. No separator.

PrettyDoc.parens — Method

Wrap a document with left and right parentheses.

PrettyDoc.pretty — Method

Using repr(o) to create a document from an object.

PrettyDoc.seg — Method

Making a document from a string. Note that including line ending characters can cause ugly pretty printing.

PrettyDoc.seplistof — Function

seplistof(separator, elements)

Create a document that represents a list of elements, separated by the given separator.

PrettyDoc.vsep — Function

Concatenates a vector of documents with the line separator " ".